Fix delta bit-packed encoding of int32 and int64
The issue is if only a single value is ever added. In that case,
the list of deltas is empty, and flushing them didn't write anything.
This is changed so that a flush is done if there are deltas
(the common case) or if exactly one value has been added. Looking
more closely into this bug, I also discovered the binary bit-packing
encoders for int32 and int64 never seemed to have worked properly
as the block isze and mini block count never got initialized.